Capstone 1 progress

+++ m1-capstone-java-cateringsystem/src/main/java/com/techelevator/service/OrderingService.java
@@ -10,18 +10,17 @@ import java.util.Map;
 
 public class OrderingService {
 
-    CustomerAccount customerAccount;
+    private final CustomerAccount customerAccount;
 
-    CateringService cateringService;
+   private final CateringService cateringService;
 
-    Cart cart;
+    private final Cart cart;
 
 
     public OrderingService(CateringService cateringService) {
         this.cateringService = cateringService;
         customerAccount= new CustomerAccount();
-        Cart cart= new Cart();
-
+        cart= new Cart();
 
     }
 
@@ -38,27 +37,61 @@ public class OrderingService {
         }
 
         Product product= inventory.get(productCode).getProduct();
-        int amountInStock= inventory.get(productCode).getAmount();
+        int amountInStock= inventory.get(productCode).getNumOfProducts();
         if(amountInStock==0){
-            return "SOLD OUT";
+            return "SOLD OUT!";
         }
         if(amountInStock<amount){
             return "Insufficient Stock!";
         }
+
+        double requiredBalanceForPurchase= product.getProductPrice()* amount;
+
+        if(customerAccount.getBalance()<requiredBalanceForPurchase){
+            return "NOT ENOUGH FUND available for purchase! Please add more money to balance.";
+        }
+
+        customerAccount.updateBalance(-1* requiredBalanceForPurchase);
+
 
         //there must be enough
         cateringService.removeFromShelf(productCode,amount);
 
         cart.addToCart(productCode,amount);
 
-        return "Successfully added to cart!";
-
-
-
-
+        return "Successfully added to cart! :)";
 
     }
 
+    public void completeTransaction(){
+        Map<String,Integer> cartItems= cart.getCartItems();
+        double totalAmount = 0;
+
+        for (String productCode : cartItems.keySet()) {
+
+            Product product = cateringService.getProductInventory().get(productCode).getProduct();
+
+            int numOfProduct = cartItems.get(productCode);
+            String productType = product.toString();
+            String productName = product.getProductName();
+            Double productPrice = product.getProductPrice();
+            Double tolCostPerProduct = numOfProduct * productPrice;
+
+            System.out.printf("%-10s %-15s %-23s %-15.2f %-15.2f\n",
+                    numOfProduct,
+                    productType,
+                    productName,
+                    productPrice,
+                    tolCostPerProduct);
+
+
+            totalAmount+=tolCostPerProduct;
+
+
+        }
+        System.out.println("Total: " + totalAmount);
+
+    }
     public CustomerAccount getCustomerAccount() {
         return customerAccount;
     }
